integration tests, kinda
