+++ tests/integration.rs
@@ -1,6 +1,7 @@
-use super::*;
-use interpreter::{Interpreter, InterpreterUtils, SupportLevel};
-use interpreters::JS_original;
+use sniprun::*;
+use sniprun::interpreter::{Interpreter, InterpreterUtils, SupportLevel};
+use sniprun::interpreters::JS_original;
+use std::sync::{Arc,Mutex};
 
 #[test]
 fn test_implements() {
